move shared contexts to contexts/common
