+++ src/renderer/contexts/common/Overlay/types.ts
