+++ src/renderer/contexts/common/Tooltip/defaults.ts
