+++ src/renderer/contexts/common/Tooltip/types.ts
