+++ src/renderer/library/Hardware/HardwareAddress/index.tsx
@@ -19,7 +19,7 @@ import type { FormEvent } from 'react';
 import type { HardwareAddressProps } from './types';
 import { getAddressChainId } from '@/renderer/Utils';
 import { useConnections } from '@/renderer/contexts/import/Connections';
-import { useTooltip } from '@/renderer/contexts/Tooltip';
+import { useTooltip } from '@/renderer/contexts/common/Tooltip';
 import { ButtonMono } from '@/renderer/kits/Buttons/ButtonMono';
 
 export const HardwareAddress = ({
