+++ src/renderer/screens/Import/Home.tsx
@@ -19,7 +19,7 @@ import { ButtonText } from '@/renderer/kits/Buttons/ButtonText';
 import { ActionItem } from '@/renderer/library/ActionItem';
 import { ModalConnectItem } from '@/renderer/kits/Overlay/structure/ModalConnectItem';
 import { ModalHardwareItem } from '@/renderer/kits/Overlay/structure/ModalHardwareItem';
-import { useHelp } from '@/renderer/contexts/Help';
+import { useHelp } from '@/renderer/contexts/common/Help';
 import type { HomeProps } from './types';
 
 export const Home = ({ setSection, setSource }: HomeProps) => {
