+++ src/renderer/screens/Import/Vault/Manage.tsx
@@ -4,7 +4,7 @@
 import { faQrcode } from '@fortawesome/free-solid-svg-icons';
 import { BodyInterfaceWrapper } from '@app/Wrappers';
 import AppSVG from '@/config/svg/ledger/polkadot.svg?react';
-import { useOverlay } from '@app/contexts/Overlay';
+import { useOverlay } from '@/renderer/contexts/common/Overlay';
 import { DragClose } from '@app/library/DragClose';
 import { ErrorBoundary } from 'react-error-boundary';
 import PolkadotVaultSVG from '@w3ux/extension-assets/PolkadotVault.svg?react';
